Testing maxflight option
This requires changing all timing internally to use floating poiont
values. I suspect this won't harm anything and may actually help in some
cases.

+++ lib/mcmd.rb
@@ -142,7 +142,7 @@ class MultipleCmd
   end
 
   def process_timeouts
-    now = Time.now.to_i
+    now = Time.now.to_f
     @subproc_by_pid.values.each do |p|
       if ((now - p.time_start) > self.perchild_timeout) and self.perchild_timeout > 0
         # expire this child process
@@ -164,7 +164,7 @@ class MultipleCmd
   end
 
   def run
-    @global_time_start = Time.now.to_i
+    @global_time_start = Time.now.to_f
     done = false
     while not done
       # start up as many as maxflight processes
@@ -227,7 +227,7 @@ class MultipleCmd
           # pid is now gone. remove from subproc_by_pid and
           # add to the processed commands list
           p = @subproc_by_pid[pid]
-          p.time_end = Time.now.to_i
+          p.time_end = Time.now.to_f
           p.retval = $?
           @subproc_by_pid.delete(pid)
           @processed_commands << p
@@ -257,7 +257,7 @@ class MultipleCmd::SubProc
 
   def initialize
     self.write_buf_position = 0
-    self.time_start = Time.now.to_i
+    self.time_start = Time.now.to_f
     self.stdout_buf = ""
     self.stderr_buf = ""
     self.terminated = false
